+++ Components/RoomEvent.js
@@ -12,16 +12,16 @@ export let RoomEvents = {
 }
 
 let eventData = [
-    (room, timestamp, fromState, toState) => {//CHECKMARK_STATE_CHANGE
+    (room, timestamp, fromState, toState) => {// CHECKMARK_STATE_CHANGE
         return timeNumber(Date.now() - timestamp) + " CHECKMARK CHANGE: " + room.checkmarkStateToName(fromState) + " -> " + room.checkmarkStateToName(toState)
     },
-    (room, timestamp, player) => {//PLAYER_ENTER
+    (room, timestamp, player) => {// PLAYER_ENTER
         return timeNumber(Date.now() - timestamp) + " ENTER: " + player.username
     },
-    (room, timestamp, player) => {//PLAYER_EXIT
+    (room, timestamp, player) => {// PLAYER_EXIT
         return timeNumber(Date.now() - timestamp) + " EXIT: " + player.username
     },
-    (room, timestamp, from, to) => {//SECRET_COUNT_CHANGE
+    (room, timestamp, from, to) => {// SECRET_COUNT_CHANGE
         return timeNumber(Date.now() - timestamp) + " SECRET CHANGE: " + from + " -> " + to
     }
 ]