+++ Render/ScoreRendering/ScoreMapRenderer.js
@@ -15,7 +15,7 @@ class ScoreMapRenderer extends MapTab {
      */
     draw(renderContext, dungeonMap, mouseX, mouseY) {
 
-        let mapData //Get map data from hotbar
+        let mapData // Get map data from hotbar
         try {
             let item = Player.getInventory().getStackInSlot(8)
             mapData = item.getItem()[m.getMapData](item.getItemStack(), World.getWorld()); // ItemStack.getItem().getMapData()
@@ -28,12 +28,12 @@ class ScoreMapRenderer extends MapTab {
 
         let { x, y, size } = renderContext.getMapDimensions()
 
-        GlStateManager.func_179094_E(); //GlStateManager.push()
+        GlStateManager.func_179094_E(); // GlStateManager.push()
         Renderer.translate(x + renderContext.borderWidth, y + renderContext.borderWidth, 1)
-        GlStateManager.func_179152_a((size - 2 * renderContext.borderWidth) / 128, (size - renderContext.borderWidth) / 128, 1); //GlStateManager.scale()
+        GlStateManager.func_179152_a((size - 2 * renderContext.borderWidth) / 128, (size - renderContext.borderWidth) / 128, 1); // GlStateManager.scale()
         GlStateManager.func_179131_c(1.0, 1.0, 1.0, 1.0); // GlStateManager.color()
         Client.getMinecraft().field_71460_t.func_147701_i().func_148250_a(mapData, true);
-        GlStateManager.func_179121_F(); //GlStateManager.pop()
+        GlStateManager.func_179121_F(); // GlStateManager.pop()
     }
 
     /**